+++ R/tar_map2.R
@@ -19,6 +19,7 @@
 #'   `command2`, and `command2` dynamically maps over
 #'   these arguments in batches.
 #' @inheritSection tar_rep Replicate-specific seeds
+#' @inheritSection tar_rep Nested futures for batched replication
 #' @return A list of new target objects.
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects