+++ R/tar_map2_count.R
@@ -16,6 +16,7 @@
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects
 #' @inheritSection tar_rep Replicate-specific seeds
+#' @inheritSection tar_rep Nested futures for batched replication
 #' @inheritParams tar_map2
 #' @inheritParams tar_map2_count_raw
 #' @examples