+++ R/tar_map2_size_raw.R
@@ -19,6 +19,7 @@
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects
 #' @inheritSection tar_rep Replicate-specific seeds
+#' @inheritSection tar_rep Nested futures for batched replication
 #' @param size Positive integer of length 1,
 #'   maximum number of rows in each batch for
 #'   the downstream (`command2`) targets. Batches