+++ NEWS.md
@@ -1,6 +1,6 @@
 # tarchetypes 0.7.2.9000
 
-
+* Support nested futures for parallelism among reps within batches (#117, @kkmann).
 
 # tarchetypes 0.7.2